requirements.txt
# Core Dependencies
openai==1.54.3
pinecone-client==5.0.0
sentence-transformers==3.2.0
mistralai==1.9.3

# Text Processing
PyPDF2==3.0.1
transformers==4.46.2
torch==2.5.1

# Web Framework
flask==3.0.3
flask-cors==5.0.0

# Utilities
tqdm==4.66.6
numpy==1.26.4
pandas==2.2.3
python-dotenv==1.0.1
requests==2.32.3
pillow==10.4.0

# ML/AI
scikit-learn==1.5.2